+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2021-01-14
+*Features*
+- `outputDatasetId` can now also use dataset name. If dataset with that name doesn't exist, a new dataset is created.
+
 ## 2020-07-10
 *Fixes:*
 - `dedup-as-loading` mode now works correctly with actor migrations. This means that this actor can finally be used for huge datasets with lower memory!